requirements-all.txt
annotated-types==0.6.0
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==42.0.2
cymem==2.0.8
fake-useragent==1.4.0
h11==0.14.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
lingua-language-detector==2.0.2
lxml==5.1.0
MarkupSafe==2.1.4
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pdfminer.six==20231228
preshed==3.0.9
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyPDF2==3.0.1
PySocks==1.7.1
python-docx==1.1.0
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
selenium==4.17.2
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.2
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
wasabi==1.1.2
weasel==0.3.4
webdriver-manager==4.0.1
wsproto==1.2.0
